Get rid of annoying compiler warnings with wrapping unused parameters with a macro. So, we can still compile cleanly with highest level of warning reports.
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%40716

+++ src/las_c_api.cpp
@@ -1411,7 +1411,7 @@ LAS_DLL LASErrorEnum LASVLR_GetData(const LASVLRH hVLR, liblas::uint8_t* data, i
     try {
         std::vector<liblas::uint8_t> *d = new std::vector<liblas::uint8_t>(((LASVLR*) hVLR)->GetData());
         *data = d->front();
-        *length = d->size();
+        *length = static_cast<int>(d->size());
     }
     catch (std::exception const& e) {
         LASError_PushError(LE_Failure, e.what(), "LASVLR_GetData");
