+++ src/lasheader.cpp
@@ -497,7 +497,7 @@ void LASHeader::SetMin(double x, double y, double z)
 void LASHeader::AddVLR(LASVLR const& v) 
 {
     m_vlrs.push_back(v);
-    m_dataRecordLen = m_vlrs.size();
+    m_dataRecordLen = static_cast<uint16_t>(m_vlrs.size());
 }
 
 LASVLR const& LASHeader::GetVLR(uint32_t index) const {
@@ -510,7 +510,7 @@ void LASHeader::DeleteVLR(uint32_t index) {
         throw std::out_of_range("index is out of range");
         
     m_vlrs.erase(m_vlrs.begin() + index);
-    m_dataRecordLen = m_vlrs.size();
+    m_dataRecordLen = static_cast<uint16_t>(m_vlrs.size());
     
 }
 void LASHeader::Init()
